Modulize StepExecutor. Update labels.
